Define the schema definition and create a validation for it in the property definition
